Make the Apple code path in Viz work with video hole quads
This is the last part of the required changes to make
SBPlayer-rendered videos show on the screen together with the rest of
the web contents.

The idea here is to allow hole punching to work by allow the videos
handled by SbPlayer to be shown as underlays: the UIViews with videos
are added before the web contents view and are thus supposed to be
shown "under" it.

On the Viz side, this is done by treating video hole quads (which are
already added by Starboard media code in `//media`) specially: when
found, these quads disable the default quads-to-CALayers code path
that normally renders each quad as a separate CALayer (with other
CALayers underneath, managed and described by `CARendererLayerTree`).
Punching a hole through this complex hierarchy of nested and sibling
CALayers is complicated and not worth the effort at the moment.

By doing so:

- `CALayerOverlayProcessor::PutForcedOverlayContentIntoUnderlays()`
  and `CALayerOverlayProcessor::PutQuadInSeparateOverlay()` are always
  called, and we use `PutQuadInSeparateOverlay()` to replace the video
  hole quad with a transparent quad just like the existing code in
  `OverlayStrategyUnderlayStarboard`.
- The entire page is received by `CARendererLayerTree` as a single
  CALayer, which includes the transparent quad created above,
  effectively making the video UIViews poke through the web contents.